+++ SGRH.Api/Controllers/ServiciosController/ServicioController.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using SGRH.Application.DTOs.Reserva;
 using SGRH.Application.DTOs.Reserva.ServicioDto;
 using SGRH.Application.Interfaces.Services;
 
@@ -39,8 +38,13 @@ namespace SGRH.Api.Controllers.ServiciosController
 
         public async Task<IActionResult> GetById(int Id) 
         {
+            if (Id <= 0)
+                return NotFound("No se Encontro el Servicio Solicitado");
+            
             var result = await _service.GetByIdAsync(Id);
 
+            if (result == null)
+                return NotFound("No se Encontro el Servicio Solicitado");
 
             return Ok(result);
         }
@@ -49,8 +53,13 @@ namespace SGRH.Api.Controllers.ServiciosController
 
         public async Task<IActionResult> Deletele(int Id, int IdUsuario) 
         {
+            if (Id <= 0 || IdUsuario <= 0)
+                return NotFound("No se Encontro el Servicio a Eliminar");
+
             var result = await _service.DeleteAsync(Id, IdUsuario);
 
+            if (result == null)
+                return NotFound("No se Encontro el Servicio a Eliminar");
 
             return Ok(result);
         }
